pyproject.toml
[tool.poetry]
name = "local agentic api"
version = "0.1.0"
description = ""
authors = ["Shannon Gray <shanngray@icloud.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.52.2"
filelock = "^3.16.1"
tavily-python = "^0.5.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
httpx = "^0.27.2"
rich = "^13.9.4"
pytz = "^2024.2"
chromadb = "^0.5.18"

[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"
pytest = "^8.3.3"
anyio = "^4.6.2.post1"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"